Fixes #62, Fixes #63

+++ src/main/java/net/wizardsoflua/lua/classes/entity/CreatureClass.java
@@ -31,7 +31,7 @@ public class CreatureClass {
     }
 
     @Override
-    public double getRotationYaw() {
+    public float getRotationYaw() {
       float v = delegate.renderYawOffset;
       return MathHelper.wrapDegrees(v);
     }
